refactor: 장바구니 관련 UI와 로직을 별도 CartView 클래스로 분리
